Refactor Compilation into feature partials
