+++ src/Raven.CodeAnalysis/Compilation.cs
@@ -4,33 +4,30 @@ using System.Collections.Immutable;
 using System.Linq;
 using System.Reflection;
 
-using Raven.CodeAnalysis.CodeGen;
 using Raven.CodeAnalysis.Symbols;
 using Raven.CodeAnalysis.Syntax;
 
 namespace Raven.CodeAnalysis;
 
-public class Compilation
+public partial class Compilation
 {
+    private readonly object _setupLock = new();
     private INamespaceSymbol? _globalNamespace;
     private readonly SyntaxTree[] _syntaxTrees;
     private readonly MetadataReference[] _references;
     internal SyntaxTree? SyntaxTreeWithFileScopedCode;
-    private readonly Dictionary<SyntaxTree, SemanticModel> _semanticModels = new Dictionary<SyntaxTree, SemanticModel>();
-    private readonly Dictionary<MetadataReference, IAssemblySymbol> _metadataReferenceSymbols = new Dictionary<MetadataReference, IAssemblySymbol>();
-    private readonly Dictionary<Assembly, IAssemblySymbol> _assemblySymbols = new Dictionary<Assembly, IAssemblySymbol>();
-    private readonly Dictionary<DelegateSignature, SynthesizedDelegateTypeSymbol> _synthesizedDelegates = new(new DelegateSignatureComparer());
-    private readonly Dictionary<SourceMethodSymbol, SynthesizedAsyncStateMachineTypeSymbol> _synthesizedAsyncStateMachines = new(ReferenceEqualityComparer.Instance);
-    private readonly Dictionary<SourceMethodSymbol, SynthesizedIteratorTypeSymbol> _synthesizedIterators = new(ReferenceEqualityComparer.Instance);
-    private int _synthesizedDelegateOrdinal;
-    private int _synthesizedAsyncStateMachineOrdinal;
-    private int _synthesizedIteratorOrdinal;
+    private readonly Dictionary<MetadataReference, IAssemblySymbol> _metadataReferenceSymbols = new();
+    private readonly Dictionary<Assembly, IAssemblySymbol> _assemblySymbols = new();
+    private readonly Dictionary<string, Assembly> _lazyMetadataAssemblies = new();
+    private MetadataLoadContext _metadataLoadContext;
+    private GlobalBinder _globalBinder;
+    private bool setup;
+    private ErrorTypeSymbol _errorTypeSymbol;
+    private NullTypeSymbol _nullTypeSymbol;
+    private UnitTypeSymbol _unitTypeSymbol;
+    private TypeResolver _typeResolver;
     private bool _sourceTypesInitialized;
     private bool _isPopulatingSourceTypes;
-    private readonly object _entryPointGate = new();
-    private bool _entryPointComputed;
-    private IMethodSymbol? _entryPoint;
-    private ImmutableArray<Diagnostic> _entryPointDiagnostics = ImmutableArray<Diagnostic>.Empty;
 
     private Compilation(string? assemblyName, SyntaxTree[] syntaxTrees, MetadataReference[] references, CompilationOptions? options = null)
     {
@@ -120,275 +117,12 @@ public class Compilation
         return new Compilation(assemblyName, _syntaxTrees, _references, Options);
     }
 
-    public SemanticModel GetSemanticModel(SyntaxTree syntaxTree)
-    {
-        EnsureSetup();
-
-        EnsureSemanticModelsCreated();
-
-        if (_semanticModels.TryGetValue(syntaxTree, out var semanticModel))
-        {
-            return semanticModel;
-        }
-
-        if (!_syntaxTrees.Contains(syntaxTree))
-        {
-            throw new ArgumentNullException(nameof(syntaxTree), "Syntax tree is not part of compilation");
-        }
-
-        semanticModel = new SemanticModel(this, syntaxTree);
-        _semanticModels[syntaxTree] = semanticModel;
-        return semanticModel;
-    }
-
-    private void EnsureSemanticModelsCreated()
-    {
-        if (_sourceTypesInitialized || _isPopulatingSourceTypes)
-            return;
-
-        try
-        {
-            _isPopulatingSourceTypes = true;
-
-            foreach (var syntaxTree in _syntaxTrees)
-            {
-                if (_semanticModels.ContainsKey(syntaxTree))
-                    continue;
-
-                var model = new SemanticModel(this, syntaxTree);
-                _semanticModels[syntaxTree] = model;
-            }
-
-            _sourceTypesInitialized = true;
-        }
-        finally
-        {
-            _isPopulatingSourceTypes = false;
-        }
-    }
-
     public MetadataReference ToMetadataReference() => new CompilationReference(this);
-
-    public EmitResult Emit(Stream peStream, Stream? pdbStream = null)
-    {
-        var diagnostics = GetDiagnostics();
-
-        if (diagnostics.Any(x => x.Severity == DiagnosticSeverity.Error))
-        {
-            return new EmitResult(false, diagnostics);
-        }
-
-        new CodeGenerator(this).Emit(peStream, pdbStream);
-
-        return new EmitResult(true, diagnostics);
-    }
-
-    public IMethodSymbol? GetEntryPoint(CancellationToken cancellationToken = default)
-    {
-        EnsureEntryPointComputed();
-        return _entryPoint;
-    }
-
-    internal ImmutableArray<Diagnostic> GetEntryPointDiagnostics(CancellationToken cancellationToken = default)
-    {
-        EnsureEntryPointComputed();
-        return _entryPointDiagnostics;
-    }
-
-    internal bool IsEntryPointCandidate(IMethodSymbol method)
-    {
-        if (method.Name != "Main" || !method.IsStatic || method.IsGenericMethod)
-            return false;
-
-        if (!method.TypeParameters.IsDefaultOrEmpty)
-            return false;
-
-        if (!IsValidEntryPointReturnType(method.ReturnType))
-            return false;
-
-        var parameters = method.Parameters;
-
-        if (parameters.Length == 0)
-            return true;
-
-        if (parameters.Length > 1)
-            return false;
-
-        var parameter = parameters[0];
-
-        if (parameter.RefKind != RefKind.None)
-            return false;
-
-        if (parameter.Type is not IArrayTypeSymbol arrayType)
-            return false;
-
-        var stringType = GetSpecialType(SpecialType.System_String);
-        return SymbolEqualityComparer.Default.Equals(arrayType.ElementType, stringType);
-    }
-
-    private bool IsValidEntryPointReturnType(ITypeSymbol returnType)
-    {
-        switch (returnType.SpecialType)
-        {
-            case SpecialType.System_Int32:
-            case SpecialType.System_Void:
-            case SpecialType.System_Unit:
-                return true;
-        }
-
-        var taskType = GetTypeByMetadataName("System.Threading.Tasks.Task");
-        if (taskType is not null && SymbolEqualityComparer.Default.Equals(returnType, taskType))
-            return true;
-
-        if (returnType is INamedTypeSymbol named && !named.IsUnboundGenericType && named.Arity == 1)
-        {
-            var taskOfT = GetTypeByMetadataName("System.Threading.Tasks.Task`1");
-            if (taskOfT is INamedTypeSymbol definition && SymbolEqualityComparer.Default.Equals(named.ConstructedFrom, definition))
-            {
-                var intType = GetSpecialType(SpecialType.System_Int32);
-                if (!named.TypeArguments.IsDefaultOrEmpty && SymbolEqualityComparer.Default.Equals(named.TypeArguments[0], intType))
-                    return true;
-            }
-        }
-
-        return false;
-    }
-
-    private void EnsureEntryPointComputed()
-    {
-        if (_entryPointComputed)
-            return;
-
-        lock (_entryPointGate)
-        {
-            if (_entryPointComputed)
-                return;
-
-            var uniqueCandidates = new Dictionary<string, IMethodSymbol>(StringComparer.Ordinal);
-
-            foreach (var method in SourceGlobalNamespace
-                .GetAllMembersRecursive()
-                .OfType<IMethodSymbol>()
-                .Where(IsEntryPointCandidate))
-            {
-                var key = method.ToDisplayString(SymbolDisplayFormat.CSharpSymbolKeyFormat);
-                if (!uniqueCandidates.ContainsKey(key))
-                    uniqueCandidates.Add(key, method);
-            }
-
-            var candidates = uniqueCandidates.Values.ToImmutableArray();
-
-            if (Options.OutputKind != OutputKind.ConsoleApplication)
-            {
-                _entryPoint = candidates.Length == 1 ? candidates[0] : null;
-                _entryPointDiagnostics = ImmutableArray<Diagnostic>.Empty;
-            }
-            else if (candidates.Length == 1)
-            {
-                _entryPoint = candidates[0];
-                _entryPointDiagnostics = ImmutableArray<Diagnostic>.Empty;
-            }
-            else if (candidates.Length > 1)
-            {
-                _entryPoint = null;
-                var builder = ImmutableArray.CreateBuilder<Diagnostic>(candidates.Length);
-
-                foreach (var candidate in candidates)
-                {
-                    var location = candidate.Locations.FirstOrDefault() ?? Location.None;
-                    builder.Add(Diagnostic.Create(CompilerDiagnostics.EntryPointIsAmbiguous, location));
-                }
-
-                _entryPointDiagnostics = builder.ToImmutable();
-            }
-            else
-            {
-                _entryPoint = null;
-                _entryPointDiagnostics = ImmutableArray<Diagnostic>.Empty;
-            }
-
-            _entryPointComputed = true;
-        }
-    }
-
-    internal INamedTypeSymbol GetMethodReferenceDelegate(IMethodSymbol methodSymbol)
-    {
-        var parameterTypes = methodSymbol.Parameters
-            .Select(p => p.Type)
-            .ToImmutableArray();
-        var refKinds = methodSymbol.Parameters
-            .Select(p => p.RefKind)
-            .ToImmutableArray();
-
-        return GetMethodReferenceDelegate(parameterTypes, refKinds, methodSymbol.ReturnType);
-    }
-
-    internal INamedTypeSymbol GetMethodReferenceDelegate(
-        ImmutableArray<ITypeSymbol> parameterTypes,
-        ImmutableArray<RefKind> refKinds,
-        ITypeSymbol returnType)
-    {
-        if (refKinds.All(static refKind => refKind == RefKind.None))
-        {
-            var functionType = CreateFunctionTypeSymbol(parameterTypes.ToArray(), returnType);
-            if (functionType is INamedTypeSymbol namedDelegate && namedDelegate.TypeKind == TypeKind.Delegate)
-                return namedDelegate;
-        }
-
-        var signature = new DelegateSignature(parameterTypes, refKinds, returnType);
-        if (_synthesizedDelegates.TryGetValue(signature, out var existing))
-            return existing;
-
-        var delegateName = $"<>f__Delegate{_synthesizedDelegateOrdinal++}";
-        var containingNamespace = SourceGlobalNamespace;
-        var synthesized = new SynthesizedDelegateTypeSymbol(
-            this,
-            delegateName,
-            parameterTypes,
-            refKinds,
-            returnType,
-            containingNamespace);
-
-        _synthesizedDelegates[signature] = synthesized;
-        return synthesized;
-    }
-
-    internal IEnumerable<INamedTypeSymbol> GetSynthesizedDelegateTypes()
-        => _synthesizedDelegates.Values;
-
-    internal SynthesizedAsyncStateMachineTypeSymbol CreateAsyncStateMachine(SourceMethodSymbol method)
-    {
-        if (_synthesizedAsyncStateMachines.TryGetValue(method, out var existing))
-            return existing;
-
-        var name = $"<>c__AsyncStateMachine{_synthesizedAsyncStateMachineOrdinal++}";
-        var stateMachine = new SynthesizedAsyncStateMachineTypeSymbol(this, method, name);
-        _synthesizedAsyncStateMachines[method] = stateMachine;
-        return stateMachine;
-    }
-
-    internal IEnumerable<SynthesizedAsyncStateMachineTypeSymbol> GetSynthesizedAsyncStateMachineTypes()
-        => _synthesizedAsyncStateMachines.Values;
-
-    internal SynthesizedIteratorTypeSymbol CreateIteratorStateMachine(SourceMethodSymbol method, IteratorMethodKind iteratorKind, ITypeSymbol elementType)
-    {
-        if (_synthesizedIterators.TryGetValue(method, out var existing))
-            return existing;
-
-        var name = $"<>c__Iterator{_synthesizedIteratorOrdinal++}";
-        var stateMachine = new SynthesizedIteratorTypeSymbol(this, method, name, iteratorKind, elementType);
-        _synthesizedIterators[method] = stateMachine;
-        return stateMachine;
-    }
-
-    internal IEnumerable<SynthesizedIteratorTypeSymbol> GetSynthesizedIteratorTypes()
-        => _synthesizedIterators.Values;
-
-    private readonly object _setupLock = new();
 
     internal void EnsureSetup()
     {
-        if (setup) return;
+        if (setup)
+            return;
 
         lock (_setupLock)
         {
@@ -403,9 +137,9 @@ public class Compilation
     private void Setup()
     {
         List<string> paths = _references
-        .OfType<PortableExecutableReference>()
-        .Select(portableExecutableReference => portableExecutableReference.FilePath)
-        .ToList();
+            .OfType<PortableExecutableReference>()
+            .Select(portableExecutableReference => portableExecutableReference.FilePath)
+            .ToList();
 
         var resolver = new PathAssemblyResolver(paths);
         _metadataLoadContext = new MetadataLoadContext(resolver);
@@ -425,32 +159,7 @@ public class Compilation
         Module = new SourceModuleSymbol(AssemblyName, (SourceAssemblySymbol)Assembly, _metadataReferenceSymbols.Values, []);
 
         SourceGlobalNamespace = (SourceNamespaceSymbol)Module.GlobalNamespace;
-
-        /*
-        foreach (var syntaxTree in SyntaxTrees)
-        {
-            var root = syntaxTree.GetRoot();
-
-            Location[] locations = [syntaxTree.GetLocation(root.Span)];
-
-            SyntaxReference[] references = [root.GetReference()];
-
-            foreach (var memberDeclaration in root.Members)
-            {
-                AnalyzeMemberDeclaration(syntaxTree, SourceGlobalNamespace, memberDeclaration);
-            }
-        }
-        */
     }
-
-    private readonly Dictionary<string, Assembly> _lazyMetadataAssemblies = new();
-    private MetadataLoadContext _metadataLoadContext;
-    private GlobalBinder _globalBinder;
-    private bool setup;
-    private ErrorTypeSymbol _errorTypeSymbol;
-    private NullTypeSymbol _nullTypeSymbol;
-    private UnitTypeSymbol _unitTypeSymbol;
-    private TypeResolver _typeResolver;
 
     internal TypeResolver TypeResolver => _typeResolver ??= new TypeResolver(this);
 
@@ -661,7 +370,7 @@ public class Compilation
                 methodDeclaration.Modifiers,
                 defaultAccessibility);
 
-            var symbol = new SourceMethodSymbol(
+            _ = new SourceMethodSymbol(
                 methodDeclaration.Identifier.ValueText, returnType,
                 ImmutableArray<SourceParameterSymbol>.Empty,
                 declaringSymbol,
@@ -673,433 +382,157 @@ public class Compilation
         }
     }
 
-    public ImmutableArray<Diagnostic> GetDiagnostics(CompilationWithAnalyzersOptions? analyzerOptions = null, CancellationToken cancellationToken = default)
+    public ITypeSymbol CreateArrayTypeSymbol(ITypeSymbol elementType, int rank = 1)
     {
-        var diagnostics = new List<Diagnostic>();
-
-        EnsureSetup();
-
-        foreach (var syntaxTree in SyntaxTrees)
-        {
-            foreach (var diagnostic in syntaxTree.GetDiagnostics(cancellationToken))
-                Add(diagnostic);
-
-            var model = GetSemanticModel(syntaxTree);
-            foreach (var diagnostic in model.GetDiagnostics(cancellationToken))
-                Add(diagnostic);
-        }
-
-        var entryPointDiagnostics = GetEntryPointDiagnostics(cancellationToken);
-        foreach (var diagnostic in entryPointDiagnostics)
-            Add(diagnostic);
-
-        if (Options.OutputKind == OutputKind.ConsoleApplication
-            && entryPointDiagnostics.IsDefaultOrEmpty
-            && GetEntryPoint(cancellationToken) is null)
-        {
-            Add(Diagnostic.Create(CompilerDiagnostics.ConsoleApplicationRequiresEntryPoint, Location.None));
-        }
-
-        return diagnostics.OrderBy(x => x.Location).ToImmutableArray();
-
-        void Add(Diagnostic diagnostic)
-        {
-            var mapped = ApplyCompilationOptions(diagnostic, analyzerOptions?.ReportSuppressedDiagnostics ?? false);
-            if (mapped is not null)
-                diagnostics.Add(mapped);
-        }
+        var ns = GlobalNamespace.LookupNamespace("System");
+        return new ArrayTypeSymbol(GetSpecialType(SpecialType.System_Array), elementType, ns, null, ns, [], rank);
     }
 
-    internal Diagnostic? ApplyCompilationOptions(Diagnostic diagnostic, bool reportSuppressedDiagnostics = false)
+    public ITypeSymbol CreateFunctionTypeSymbol(ITypeSymbol[] parameterTypes, ITypeSymbol returnType)
     {
-        if (Options.SpecificDiagnosticOptions.TryGetValue(diagnostic.Descriptor.Id, out var report))
-        {
-            if (report == ReportDiagnostic.Suppress)
-                return reportSuppressedDiagnostics ? diagnostic.WithSuppression(true) : null;
+        var systemNamespace = GlobalNamespace.LookupNamespace("System");
 
-            if (report != ReportDiagnostic.Default)
-            {
-                var severity = report switch
-                {
-                    ReportDiagnostic.Error => DiagnosticSeverity.Error,
-                    ReportDiagnostic.Warn => DiagnosticSeverity.Warning,
-                    ReportDiagnostic.Info => DiagnosticSeverity.Info,
-                    ReportDiagnostic.Hidden => DiagnosticSeverity.Hidden,
-                    _ => diagnostic.Severity
-                };
+        var allTypes = parameterTypes.ToList();
+        bool isAction = returnType.SpecialType == SpecialType.System_Void || returnType.SpecialType == SpecialType.System_Unit;
 
-                if (severity != diagnostic.Severity)
-                    return diagnostic.WithSeverity(severity);
-            }
-        }
+        if (!isAction)
+            allTypes.Add(returnType);
 
-        return diagnostic;
+        string delegateName = isAction ? "Action" : "Func";
+
+        var delegateType = systemNamespace?.GetMembers(delegateName)
+            .OfType<INamedTypeSymbol>()
+            .FirstOrDefault(t => t.Arity == allTypes.Count);
+
+        if (delegateType is null)
+            return ErrorTypeSymbol;
+
+        return delegateType.Construct(allTypes.ToArray());
     }
 
-    public Conversion ClassifyConversion(ITypeSymbol source, ITypeSymbol destination)
+    public ITypeSymbol CreateTupleTypeSymbol(IEnumerable<(string? name, ITypeSymbol type)> elements)
     {
-        static ITypeSymbol Unalias(ITypeSymbol type, ref bool wasAlias)
+        var systemNamespace = GlobalNamespace.LookupNamespace("System");
+        var elementArray = elements.ToArray();
+
+        var tupleDefinition = systemNamespace?.GetMembers("ValueTuple")
+            .OfType<INamedTypeSymbol>()
+            .FirstOrDefault(t => t.Arity == elementArray.Length);
+
+        if (tupleDefinition is null)
+            return ErrorTypeSymbol;
+
+        var underlying = (INamedTypeSymbol)tupleDefinition.Construct(elementArray.Select(e => e.type).ToArray());
+        var tuple = new TupleTypeSymbol(underlying, null, null, null, []);
+
+        var fields = new List<IFieldSymbol>();
+        int i = 0;
+        foreach (var tupleField in underlying.GetMembers().OfType<SubstitutedFieldSymbol>())
         {
-            while (true)
+            var name = elementArray[i].name ?? $"Item{i + 1}";
+            fields.Add(new TupleFieldSymbol(name, tupleField, underlying, []));
+            i++;
+        }
+
+        tuple.SetTupleElements(fields);
+
+        return tuple;
+    }
+
+    public ITypeSymbol ConstructGenericType(INamedTypeSymbol genericDefinition, ITypeSymbol[] typeArgs)
+    {
+        return genericDefinition.Construct(typeArgs);
+    }
+
+    public ITypeSymbol ResolvePredefinedType(PredefinedTypeSyntax predefinedType)
+    {
+        var keywordKind = predefinedType.Keyword.Kind;
+
+        var specialType = keywordKind switch
+        {
+            SyntaxKind.BoolKeyword => SpecialType.System_Boolean,
+            SyntaxKind.CharKeyword => SpecialType.System_Char,
+            SyntaxKind.DoubleKeyword => SpecialType.System_Double,
+            SyntaxKind.IntKeyword => SpecialType.System_Int32,
+            SyntaxKind.ObjectKeyword => SpecialType.System_Object,
+            SyntaxKind.StringKeyword => SpecialType.System_String,
+            SyntaxKind.UnitKeyword => SpecialType.System_Unit,
+            _ => throw new Exception($"Unexpected predefined keyword: {keywordKind}")
+        };
+
+        return GetSpecialType(specialType)
+               ?? throw new Exception($"Special type not found for: {specialType}");
+    }
+
+    public ITypeSymbol? GetType(Type type)
+    {
+        return TypeResolver.ResolveType(type);
+    }
+
+    public ISymbol? GetAssemblyOrModuleSymbol(MetadataReference metadataReference)
+    {
+        if (!_metadataReferenceSymbols.TryGetValue(metadataReference, out var symbol))
+        {
+            switch (metadataReference)
             {
-                if (!type.IsAlias)
-                    return type;
-
-                wasAlias = true;
-
-                if (type.UnderlyingSymbol is ITypeSymbol aliasTarget)
+                case PortableExecutableReference per:
                 {
-                    type = aliasTarget;
-                    continue;
-                }
-
-                return type;
-            }
-        }
-
-        bool sourceUsedAlias = false;
-        source = Unalias(source, ref sourceUsedAlias);
-
-        bool destinationUsedAlias = false;
-        destination = Unalias(destination, ref destinationUsedAlias);
-
-        // Temporary
-        if (destination is null) return Conversion.None;
-
-        var aliasInvolved = sourceUsedAlias || destinationUsedAlias;
-
-        Conversion Finalize(Conversion conversion)
-        {
-            if (!conversion.Exists)
-                return conversion;
-
-            return conversion.WithAlias(aliasInvolved);
-        }
-
-        if (source is LiteralTypeSymbol litSrc && destination is LiteralTypeSymbol litDest)
-            return Equals(litSrc.ConstantValue, litDest.ConstantValue)
-                ? Finalize(new Conversion(isImplicit: true, isIdentity: true))
-                : Conversion.None;
-
-        if (destination is LiteralTypeSymbol)
-            return Conversion.None;
-
-        if (SymbolEqualityComparer.Default.Equals(source, destination) &&
-            source is not NullableTypeSymbol &&
-            destination is not NullableTypeSymbol)
-        {
-            // Identity conversion
-            return Finalize(new Conversion(isImplicit: true, isIdentity: true));
-        }
-
-        static bool UnionContainsNull(IUnionTypeSymbol union)
-        {
-            foreach (var member in union.Types)
-            {
-                if (member.TypeKind == TypeKind.Null)
-                    return true;
-
-                if (member is IUnionTypeSymbol nested && UnionContainsNull(nested))
-                    return true;
-            }
-
-            return false;
-        }
-
-        if (source.TypeKind == TypeKind.Null)
-        {
-            if (destination.TypeKind == TypeKind.Nullable)
-                return Finalize(new Conversion(isImplicit: true, isReference: true));
-
-            if (destination is IUnionTypeSymbol unionDest && UnionContainsNull(unionDest))
-                return Finalize(new Conversion(isImplicit: true, isReference: true));
-
-            return Conversion.None;
-        }
-
-        if (source is NullableTypeSymbol nullableSource)
-        {
-            var conv = ClassifyConversion(nullableSource.UnderlyingType, destination);
-            if (conv.Exists)
-            {
-                var isImplicit = !SymbolEqualityComparer.Default.Equals(nullableSource.UnderlyingType, destination) && conv.IsImplicit;
-                return Finalize(new Conversion(
-                    isImplicit: isImplicit,
-                    isIdentity: conv.IsIdentity,
-                    isNumeric: conv.IsNumeric,
-                    isReference: conv.IsReference,
-                    isBoxing: conv.IsBoxing,
-                    isUnboxing: conv.IsUnboxing,
-                    isUserDefined: conv.IsUserDefined,
-                    isAlias: conv.IsAlias,
-                    methodSymbol: conv.MethodSymbol));
-            }
-        }
-
-        if (destination is NullableTypeSymbol nullableDest)
-        {
-            if (source is IUnionTypeSymbol unionSource &&
-                unionSource.Types.Count() == 2 &&
-                unionSource.Types.Any(t => t.TypeKind == TypeKind.Null) &&
-                unionSource.Types.Any(t => SymbolEqualityComparer.Default.Equals(t, nullableDest.UnderlyingType)))
-            {
-                return Finalize(new Conversion(isImplicit: true, isReference: true));
-            }
-
-            var conv = ClassifyConversion(source, nullableDest.UnderlyingType);
-            if (conv.Exists)
-                return Finalize(new Conversion(
-                    isImplicit: true,
-                    isIdentity: false,
-                    isNumeric: conv.IsNumeric,
-                    isReference: conv.IsReference || !source.IsValueType,
-                    isBoxing: conv.IsBoxing,
-                    isUnboxing: conv.IsUnboxing,
-                    isUserDefined: conv.IsUserDefined,
-                    isAlias: conv.IsAlias,
-                    methodSymbol: conv.MethodSymbol));
-        }
-
-        if (source is IUnionTypeSymbol unionSource2)
-        {
-            var conversions = unionSource2.Types.Select(t => ClassifyConversion(t, destination)).ToArray();
-            if (conversions.All(c => c.Exists))
-            {
-                var isImplicit = conversions.All(c => c.IsImplicit);
-                var isAlias = conversions.Any(c => c.IsAlias);
-                var destinationIsValueType = destination.IsValueType;
-
-                return Finalize(new Conversion(
-                    isImplicit: isImplicit,
-                    isReference: !destinationIsValueType,
-                    isUnboxing: destinationIsValueType,
-                    isAlias: isAlias));
-            }
-
-            return Conversion.None;
-        }
-
-        if (source.SpecialType == SpecialType.System_Void)
-            return Conversion.None;
-
-        var objType = GetSpecialType(SpecialType.System_Object);
-
-        if (destination.Equals(objType, SymbolEqualityComparer.Default))
-        {
-            if (source.Equals(objType, SymbolEqualityComparer.Default))
-            {
-                return Conversion.None;
-            }
-
-            // Assigning to object
-            return Finalize(new Conversion(
-                isImplicit: true,
-                isReference: !source.IsValueType,
-                isBoxing: source.IsValueType));
-        }
-
-        if (destination is IUnionTypeSymbol unionType)
-        {
-            Conversion matchConversion = default;
-            var foundMatch = false;
-
-            foreach (var branch in unionType.Types)
-            {
-                var branchConversion = ClassifyConversion(source, branch);
-                if (branchConversion.Exists)
-                {
-                    matchConversion = branchConversion;
-                    foundMatch = true;
+                    var assembly = _metadataLoadContext.LoadFromAssemblyPath(per.FilePath);
+                    symbol = GetAssembly(assembly);
                     break;
                 }
-            }
-
-            if (!foundMatch)
-                return Conversion.None;
-
-            return Finalize(new Conversion(
-                isImplicit: true,
-                isBoxing: source.IsValueType,
-                isAlias: matchConversion.IsAlias));
-        }
-
-        if (source is LiteralTypeSymbol litSrc2)
-            return Finalize(ClassifyConversion(litSrc2.UnderlyingType, destination));
-
-        if (IsReferenceConversion(source, destination))
-        {
-            // Reference conversion
-            return Finalize(new Conversion(isImplicit: true, isReference: true));
-        }
-
-        if (IsExplicitReferenceConversion(source, destination))
-        {
-            // Explicit reference conversion (e.g., downcasts)
-            return Finalize(new Conversion(isImplicit: false, isReference: true));
-        }
-
-        if (IsBoxingConversion(source, destination))
-        {
-            // Boxing conversion
-            return Finalize(new Conversion(isImplicit: true, isBoxing: true));
-        }
-
-        if (IsUnboxingConversion(source, destination))
-        {
-            // Unboxing conversion
-            return Finalize(new Conversion(isImplicit: false, isUnboxing: true));
-        }
-
-        if (IsImplicitNumericConversion(source, destination))
-        {
-            // Implicit numeric conversion
-            return Finalize(new Conversion(isImplicit: true, isNumeric: true));
-        }
-
-        if (IsExplicitNumericConversion(source, destination))
-        {
-            // Explicit numeric conversion
-            return Finalize(new Conversion(isImplicit: false, isNumeric: true));
-        }
-
-        // User-defined conversions
-        var sourceNamed = source as INamedTypeSymbol;
-        var destinationNamed = destination as INamedTypeSymbol;
-
-        if (sourceNamed != null || destinationNamed != null)
-        {
-            IEnumerable<IMethodSymbol> candidateConversions =
-                Enumerable.Empty<IMethodSymbol>();
-
-            if (sourceNamed != null)
-                candidateConversions = candidateConversions.Concat(sourceNamed.GetMembers().OfType<IMethodSymbol>());
-            if (destinationNamed != null && !SymbolEqualityComparer.Default.Equals(source, destination))
-                candidateConversions = candidateConversions.Concat(destinationNamed.GetMembers().OfType<IMethodSymbol>());
-
-            foreach (var method in candidateConversions)
-            {
-                if (method.MethodKind is MethodKind.Conversion &&
-                    method.Parameters.Length == 1 &&
-                    SymbolEqualityComparer.Default.Equals(method.Parameters[0].Type, source) &&
-                    SymbolEqualityComparer.Default.Equals(method.ReturnType, destination))
+                case CompilationReference cr:
                 {
-                    var isImplicit = method.Name == "op_Implicit";
-                    return Finalize(new Conversion(isImplicit: isImplicit, isUserDefined: true, methodSymbol: method));
+                    var compilation = cr.Compilation;
+                    compilation.EnsureSetup();
+                    symbol = compilation.Assembly;
+                    break;
                 }
-            }
-        }
-
-        // No valid conversion
-        return Conversion.None;
-    }
-
-    private bool IsReferenceConversion(ITypeSymbol source, ITypeSymbol destination)
-    {
-        // Must both be reference types
-        if (source.IsValueType || destination.IsValueType)
-            return false;
-
-        // Identity conversion is not a reference conversion
-        if (SymbolEqualityComparer.Default.Equals(source, destination))
-            return false;
-
-        // Walk base types to see if destination is in the chain
-        var current = source.BaseType;
-        while (current is not null)
-        {
-            if (SymbolEqualityComparer.Default.Equals(current, destination))
-                return true;
-
-            current = current.BaseType;
-        }
-
-        if (destination is INamedTypeSymbol destinationNamed &&
-            destinationNamed.TypeKind == TypeKind.Interface)
-        {
-            if (SemanticFacts.ImplementsInterface(source, destinationNamed, SymbolEqualityComparer.Default))
-                return true;
-        }
-
-        if (source.TypeKind == TypeKind.Interface && destination.SpecialType is SpecialType.System_Object)
-            return true;
-
-        return false;
-    }
-
-    private bool IsExplicitReferenceConversion(ITypeSymbol source, ITypeSymbol destination)
-    {
-        if (source.IsValueType || destination.IsValueType)
-            return false;
-
-        if (SymbolEqualityComparer.Default.Equals(source, destination))
-            return false;
-
-        var comparer = SymbolEqualityComparer.Default;
-
-        if (SemanticFacts.IsDerivedFrom(destination, source, comparer))
-            return true;
-
-        if (source.SpecialType is SpecialType.System_Object && !destination.IsValueType)
-            return true;
-
-        if (source is INamedTypeSymbol sourceInterface && sourceInterface.TypeKind == TypeKind.Interface)
-        {
-            if (destination.SpecialType is SpecialType.System_Object)
-                return true;
-
-            if (destination is INamedTypeSymbol destinationNamed &&
-                destinationNamed.TypeKind != TypeKind.Interface &&
-                SemanticFacts.ImplementsInterface(destinationNamed, sourceInterface, comparer))
-            {
-                return true;
+                default:
+                    throw new InvalidOperationException();
             }
 
-            if (destination is INamedTypeSymbol sourceToTargetInterface && sourceToTargetInterface.TypeKind == TypeKind.Interface &&
-                (SemanticFacts.ImplementsInterface(sourceInterface, sourceToTargetInterface, comparer) ||
-                 SemanticFacts.ImplementsInterface(sourceToTargetInterface, sourceInterface, comparer)))
-            {
-                return true;
-            }
+            _metadataReferenceSymbols[metadataReference] = (IAssemblySymbol)symbol!;
         }
+        return symbol;
+    }
 
-        if (destination is INamedTypeSymbol targetInterface && targetInterface.TypeKind == TypeKind.Interface &&
-            SemanticFacts.ImplementsInterface(source, targetInterface, comparer))
+    private IAssemblySymbol GetAssembly(Assembly assembly)
+    {
+        if (_assemblySymbols.TryGetValue(assembly, out var asss))
         {
-            return true;
+            return asss;
         }
 
-        return false;
-    }
+        PEAssemblySymbol assemblySymbol = new PEAssemblySymbol(assembly, []);
 
-    private bool IsBoxingConversion(ITypeSymbol source, ITypeSymbol destination)
-    {
-        return source.IsValueType && destination.SpecialType is SpecialType.System_Object;
-    }
+        var refs = assembly.GetReferencedAssemblies();
 
-    private bool IsUnboxingConversion(ITypeSymbol source, ITypeSymbol destination)
-    {
-        return source.SpecialType is SpecialType.System_Object && destination.IsValueType;
-    }
+        assemblySymbol.AddModules(
+            new PEModuleSymbol(
+                TypeResolver,
+                assemblySymbol,
+                assembly.ManifestModule,
+                [],
+                refs.Select(x =>
+                {
+                    try
+                    {
+                        var loadedAssembly = _metadataLoadContext.LoadFromAssemblyName(x);
+                        if (loadedAssembly is null)
+                            return null;
 
-    private bool IsImplicitNumericConversion(ITypeSymbol source, ITypeSymbol destination)
-    {
-        // Add logic for implicit numeric conversions (e.g., int to long, float to double)
-        var sourceType = source.SpecialType;
-        var destType = destination.SpecialType;
+                        return GetAssembly(loadedAssembly);
+                    }
+                    catch
+                    {
+                        return null;
+                    }
+                }).Where(x => x is not null).ToArray()));
 
-        return (sourceType is SpecialType.System_Int32 && destType is SpecialType.System_Int64) ||
-               (sourceType is SpecialType.System_Int32 && destType is SpecialType.System_Double) ||
-               (sourceType is SpecialType.System_Single && destType is SpecialType.System_Double);
-    }
+        _assemblySymbols[assembly] = assemblySymbol;
 
-    private bool IsExplicitNumericConversion(ITypeSymbol source, ITypeSymbol destination)
-    {
-        // Add logic for explicit numeric conversions (e.g., double to int)
-        var sourceType = source.SpecialType;
-        var destType = destination.SpecialType;
-
-        return (sourceType is SpecialType.System_Double && destType is SpecialType.System_Int32) ||
-               (sourceType is SpecialType.System_Int64 && destType is SpecialType.System_Int32);
+        return assemblySymbol;
     }
 
     public INamedTypeSymbol? GetTypeByMetadataName(string metadataName)
@@ -1209,235 +642,6 @@ public class Compilation
 
         var type = GetTypeByMetadataName(metadataName);
         return type ?? (INamedTypeSymbol)ErrorTypeSymbol;
-    }
-
-    public ITypeSymbol ResolvePredefinedType(PredefinedTypeSyntax predefinedType)
-    {
-        var keywordKind = predefinedType.Keyword.Kind;
-
-        var specialType = keywordKind switch
-        {
-            SyntaxKind.BoolKeyword => SpecialType.System_Boolean,
-            //SyntaxKind.ByteKeyword => SpecialType.System_Byte,
-            SyntaxKind.CharKeyword => SpecialType.System_Char,
-            //SyntaxKind.DecimalKeyword => SpecialType.System_Decimal,
-            SyntaxKind.DoubleKeyword => SpecialType.System_Double,
-            //SyntaxKind.FloatKeyword => SpecialType.System_Single,
-            SyntaxKind.IntKeyword => SpecialType.System_Int32,
-            //SyntaxKind.LongKeyword => SpecialType.System_Int64,
-            SyntaxKind.ObjectKeyword => SpecialType.System_Object,
-            //SyntaxKind.SByteKeyword => SpecialType.System_SByte,
-            //SyntaxKind.ShortKeyword => SpecialType.System_Int16,
-            SyntaxKind.StringKeyword => SpecialType.System_String,
-            //SyntaxKind.UIntKeyword => SpecialType.System_UInt32,
-            //SyntaxKind.ULongKeyword => SpecialType.System_UInt64,
-            //SyntaxKind.UShortKeyword => SpecialType.System_UInt16,
-            SyntaxKind.UnitKeyword => SpecialType.System_Unit,
-            _ => throw new Exception($"Unexpected predefined keyword: {keywordKind}")
-        };
-
-        return GetSpecialType(specialType)
-               ?? throw new Exception($"Special type not found for: {specialType}");
-    }
-
-    public ITypeSymbol? GetType(Type type)
-    {
-        return TypeResolver.ResolveType(type);
-    }
-
-    public ISymbol? GetAssemblyOrModuleSymbol(MetadataReference metadataReference)
-    {
-        //if (!_references.Contains(metadataReference))
-        //    throw new InvalidOperationException();
-
-        if (!_metadataReferenceSymbols.TryGetValue(metadataReference, out var symbol))
-        {
-            switch (metadataReference)
-            {
-                case PortableExecutableReference per:
-                    {
-                        var assembly = _metadataLoadContext.LoadFromAssemblyPath(per.FilePath);
-                        symbol = GetAssembly(assembly);
-                        break;
-                    }
-                case CompilationReference cr:
-                    {
-                        var compilation = cr.Compilation;
-                        compilation.EnsureSetup();
-                        symbol = compilation.Assembly;
-                        break;
-                    }
-                default:
-                    throw new InvalidOperationException();
-            }
-
-            _metadataReferenceSymbols[metadataReference] = (IAssemblySymbol)symbol!;
-        }
-        return symbol;
-    }
-
-    private IAssemblySymbol GetAssembly(Assembly assembly)
-    {
-        if (_assemblySymbols.TryGetValue(assembly, out var asss))
-        {
-            return asss;
-        }
-
-        PEAssemblySymbol assemblySymbol = new PEAssemblySymbol(assembly, []);
-
-        var refs = assembly.GetReferencedAssemblies();
-
-        assemblySymbol.AddModules(
-            new PEModuleSymbol(
-                TypeResolver,
-                assemblySymbol,
-                assembly.ManifestModule,
-                [],
-                refs.Select(x =>
-                {
-                    try
-                    {
-                        var loadedAssembly = _metadataLoadContext.LoadFromAssemblyName(x);
-                        if (loadedAssembly is null)
-                            return null;
-
-                        return GetAssembly(loadedAssembly);
-                    }
-                    catch
-                    {
-                        // Ignore failed loads
-                        return null;
-                    }
-                }).Where(x => x is not null).ToArray()));
-
-        _assemblySymbols[assembly] = assemblySymbol;
-
-        return assemblySymbol;
-    }
-
-    public ITypeSymbol CreateArrayTypeSymbol(ITypeSymbol elementType, int rank = 1)
-    {
-        var ns = GlobalNamespace.LookupNamespace("System");
-        return new ArrayTypeSymbol(GetSpecialType(SpecialType.System_Array), elementType, ns, null, ns, [], rank);
-    }
-
-    /*
-    internal object CreateDelegateTypeSymbol(List<ITypeSymbol> parameterTypes, ITypeSymbol returnType)
-    {
-        var ns = GlobalNamespace.LookupNamespace("System");
-        return new DelegateTypeSymbol(GetSpecialType(SpecialType.System_Delegate), returnType, parameterTypes, ns, null, ns, []);
-    }
-    */
-
-    public ITypeSymbol CreateFunctionTypeSymbol(ITypeSymbol[] parameterTypes, ITypeSymbol returnType)
-    {
-        var systemNamespace = GlobalNamespace.LookupNamespace("System");
-
-        var allTypes = parameterTypes.ToList();
-        bool isAction = returnType.SpecialType == SpecialType.System_Void || returnType.SpecialType == SpecialType.System_Unit;
-
-        if (!isAction)
-            allTypes.Add(returnType);
-
-        string delegateName = isAction ? "Action" : "Func";
-
-        var delegateType = systemNamespace?.GetMembers(delegateName)
-            .OfType<INamedTypeSymbol>()
-            .FirstOrDefault(t => t.Arity == allTypes.Count);
-
-        if (delegateType is null)
-            return ErrorTypeSymbol;
-
-        // Construct the generic Func<> or Action<> with type arguments
-        return delegateType.Construct(allTypes.ToArray());
-    }
-
-    public ITypeSymbol CreateTupleTypeSymbol(IEnumerable<(string? name, ITypeSymbol type)> elements)
-    {
-        var systemNamespace = GlobalNamespace.LookupNamespace("System");
-        var elementArray = elements.ToArray();
-
-        var tupleDefinition = systemNamespace?.GetMembers("ValueTuple")
-            .OfType<INamedTypeSymbol>()
-            .FirstOrDefault(t => t.Arity == elementArray.Length);
-
-        if (tupleDefinition is null)
-            return ErrorTypeSymbol;
-
-        var underlying = (INamedTypeSymbol)tupleDefinition.Construct(elementArray.Select(e => e.type).ToArray());
-        var tuple = new TupleTypeSymbol(underlying, null, null, null, []);
-
-        var fields = new List<IFieldSymbol>();
-        int i = 0;
-        foreach (var tupleField in underlying.GetMembers().OfType<SubstitutedFieldSymbol>())
-        {
-            var name = elementArray[i].name ?? $"Item{i + 1}";
-            fields.Add(new TupleFieldSymbol(name, tupleField, underlying, []));
-            i++;
-        }
-
-        tuple.SetTupleElements(fields);
-
-        return tuple;
-    }
-
-    public ITypeSymbol ConstructGenericType(INamedTypeSymbol genericDefinition, ITypeSymbol[] typeArgs)
-    {
-        return genericDefinition.Construct(typeArgs);
-    }
-
-    private readonly struct DelegateSignature
-    {
-        public DelegateSignature(ImmutableArray<ITypeSymbol> parameterTypes, ImmutableArray<RefKind> refKinds, ITypeSymbol returnType)
-        {
-            ParameterTypes = parameterTypes;
-            RefKinds = refKinds;
-            ReturnType = returnType;
-        }
-
-        public ImmutableArray<ITypeSymbol> ParameterTypes { get; }
-
-        public ImmutableArray<RefKind> RefKinds { get; }
-
-        public ITypeSymbol ReturnType { get; }
-    }
-
-    private sealed class DelegateSignatureComparer : IEqualityComparer<DelegateSignature>
-    {
-        public bool Equals(DelegateSignature x, DelegateSignature y)
-        {
-            if (!SymbolEqualityComparer.Default.Equals(x.ReturnType, y.ReturnType))
-                return false;
-
-            if (x.ParameterTypes.Length != y.ParameterTypes.Length || x.RefKinds.Length != y.RefKinds.Length)
-                return false;
-
-            for (var i = 0; i < x.ParameterTypes.Length; i++)
-            {
-                if (x.RefKinds[i] != y.RefKinds[i])
-                    return false;
-
-                if (!SymbolEqualityComparer.Default.Equals(x.ParameterTypes[i], y.ParameterTypes[i]))
-                    return false;
-            }
-
-            return true;
-        }
-
-        public int GetHashCode(DelegateSignature obj)
-        {
-            var hash = SymbolEqualityComparer.Default.GetHashCode(obj.ReturnType);
-
-            for (var i = 0; i < obj.ParameterTypes.Length; i++)
-            {
-                hash = HashCode.Combine(
-                    hash,
-                    SymbolEqualityComparer.Default.GetHashCode(obj.ParameterTypes[i]),
-                    obj.RefKinds[i]);
-            }
-
-            return hash;
-        }
     }
 
     private static void InitializeTypeParameters(SourceNamedTypeSymbol typeSymbol, TypeParameterListSyntax? typeParameterList, SyntaxTree syntaxTree)
